testing in progress

+++ src/convolution.cuh
@@ -9,8 +9,8 @@ namespace kernel {
 
 template<typename T>
 __global__ void im2row(T *im, T *output,
-        size_t imageWidth, size_t imageHeight,
-        size_t filterSize, size_t Din, size_t stride, size_t padding);
+        int imageWidth, int imageHeight,
+        int filterSize, int Din, int stride, int padding);
 
 }
 
@@ -18,8 +18,8 @@ namespace gpu {
 
 template<typename T>
 void im2row(DeviceBuffer<T> &im, DeviceBuffer<T> &output,
-        size_t imageWidth, size_t imageHeight,
-        size_t filterSize, size_t Din, size_t stride, size_t padding);
+        int imageWidth, int imageHeight,
+        int filterSize, int Din, int stride, int padding);
 
 }
